+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 2.2.0 (25-Jul-2025)
 - Improved the GlobalProtect VPN IP detection logic
+- Added an option to show if an app is installed (Feature Request #18; thanks, @ScottEKendall!)
 
 ### 2.1.0 (24-Jul-2025)
 - Added an `operationMode` of "debug" to specifically enable swiftDialog debugging